test: refactored the model card patterns to be loaded from a fixture
